Move menu shader code over to menu_driver.c
